backend_rework/requirements.txt
flask
psycopg2-binary
flask-cors